修复：时区导致今日 token 总量不变，全面维护前后端代码
- 修复 summary/cost/db 中 UTC→Local 时区对齐智谱 API
- summary/cost 改为 async 并行 API 调用，消除 UI 卡顿
- notifier 释放 DB 锁后再发通知，避免阻塞
- add_account 先写 DB 再存凭据，防止孤立凭据
- set_primary_account 验证 id 存在性
- crypto Mutex poisoned 时恢复而非静默失败
- total_tokens_24h 改为 REAL 类型避免精度截断
- Windows 托盘图标 100% 时动态缩放
- 前端：条件渲染、防抖刷新、input 状态修复、semver 比较

+++ src-tauri/src/commands/history.rs
@@ -51,7 +51,7 @@ pub struct TokenHistoryPoint {
     pub timestamp: String,
     pub token_pct: f64,
     pub time_pct: f64,
-    pub tokens_24h: Option<i64>,
+    pub tokens_24h: Option<f64>,
 }
 
 #[tauri::command]
